requirements.txt
anthropic
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
notion-client
chromadb